agnostic: haunted can use lit-html or lighterhtml
this commit adds 'configureComponent' and 'configureVirtual' functions
which allow users to choose which rendering engine to use

now we can use lighterhtml, or different versions of lit-html

also adds "src/lighter-component.js", which allows out-of-the-box
usage of lighterhtml

no breaking changes have been made, only additions

no additions have been made to the topology of "haunted.js",
the new features must be explicitly imported, eg:

    import {lighterComponent} from "haunted/src/lighter-component.js"

SOURCE CODE CHANGES:

   *new*
  [src/adapt-lighter-html.js]
   - a function which adapts the lighterhtml render function to
     match the signature of the lit-html render function
   - note: the new 'configureComponent' function only accepts a render
     function which matches the signature of lit-html's

  [src/component.js]
   - preconfigured lit-html component function

   *new*
  [src/configure-component.js]
   - function which accepts a 'render' function and returns a
     component function for creating components

   *new*
  [src/configure-virtual.js]
   - function which accepts lit-html's 'directive' function and
     returns a 'virtual' function for creating virtual components

  [src/core.js]
   - now uncoupled from lit-html
   - 'Container' constructor now accepts 'render' function

  [src/haunted.js]
   - re-export lit-html functions from "lit.js" instead of "core.js"

   *new*
  [src/lighter-component.js]
   - preconfigured lighterhtml component function

  [src/virtual.js]
   - preconfigured lit-html virtual component function

TYPINGS CHANGES:

  [index.d.ts]
   - add signatures for 'configureComponent' and 'configureVirtual'

  [test/types-test.ts]
   - tests for 'configureComponent' and 'configureVirtual'

EXAMPLE CHANGES:

   *new*
  [examples/counter-lighter.html]
   - usage demo for 'lighterComponent'

ADD LIGHTERHTML RE-EXPORT:

   *new*
  [lighterhtml.js]
   - re-export from unpkg
   - did not install as dependency or involve symlink,
     to avoid increasing bundle size

README CHANGES:

  [readme.md]
   - add section "Configure component function [...]"
     to explain how to swap lit-html versions or use lighterhtml
