+++ lighterhtml.js
@@ -1,0 +1,2 @@
+
+export { render, html } from 'https://unpkg.com/lighterhtml@^0.9.3/esm/index.js?module';
