+++ src/adapt-lighter-html.js
@@ -1,0 +1,3 @@
+
+export const adaptLighterHtml = render =>
+  (result, node) => render(node, () => result);
