+++ src/configure-virtual.js
@@ -1,0 +1,72 @@
+import { Container } from './core.js';
+
+const partToContainer = new WeakMap();
+const containerToPart = new WeakMap();
+
+class DirectiveContainer extends Container {
+  constructor(render, renderer, part) {
+    super(render, renderer, part);
+    this.virtual = true;
+  }
+
+  commit(result) {
+    this.host.setValue(result);
+    this.host.commit();
+  }
+
+  teardown() {
+    super.teardown();
+    let part = containerToPart.get(this);
+    partToContainer.delete(part);
+  }
+}
+
+
+
+function agnosticVirtual(directive, render, renderer) {
+  function factory(...args) {
+    return part => {
+      let cont = partToContainer.get(part);
+      if(!cont) {
+        cont = new DirectiveContainer(render, renderer, part);
+        partToContainer.set(part, cont);
+        containerToPart.set(cont, part);
+        teardownOnRemove(cont, part);
+      }
+      cont.args = args;
+      cont.update();
+    };
+  }
+
+  return directive(factory);
+}
+
+const includes = Array.prototype.includes;
+
+function teardownOnRemove(cont, part, node = part.startNode) {
+  let frag = node.parentNode;
+  let mo = new MutationObserver(mutations => {
+    for(let mutation of mutations) {
+      if(includes.call(mutation.removedNodes, node)) {
+        mo.disconnect();
+
+        if(node.parentNode instanceof ShadowRoot) {
+          teardownOnRemove(cont, part);
+        } else {
+          cont.teardown();
+        }
+        break;
+      } else if(includes.call(mutation.addedNodes, node.nextSibling)) {
+        mo.disconnect();
+        teardownOnRemove(cont, part, node.nextSibling);
+        break;
+      }
+    }
+  });
+  mo.observe(frag, { childList: true });
+}
+
+const configureVirtual = ({directive, render}) =>
+  renderer => agnosticVirtual(directive, render, renderer);
+
+export { configureVirtual };
