+++ src/haunted.js
@@ -1,5 +1,5 @@
 export { component } from './component.js';
-export { html, render } from './core.js';
+export { html, render } from './lit.js';
 export { useCallback } from "./use-callback.js";
 export { useEffect } from './use-effect.js';
 export { useState } from './use-state.js';
